Added guide, batch funcs
- Added batch operations: MoveOutterBorders, MoveInnerBorders
- Added a user guide and important feature's summary

I'm surprised to still find important features to add after about 5 years! This is a textbook case of "evolution by real use": it was impossible to change the distance of the borders (outer and inner) without doing it side by side and without having to validate at each step. Now it's done, we can quickly move several borders simultaneously, or all from the same value, in a single operation (a single validation after the calculations). This is not just a convenience feature; it's a deep, low-level big babidi boost that also improves the clarity of the API.