Important retro-update
When I added MoveOuterBorders and MoveInnerBorders, I noticed that the Margins properties weren't in the code. After some investigation, I discovered I'd adapted the code from an older, incomplete version.

Fortunately, I was able to find the 2021 version, which I left as the most complete. In it, I aligned with the new nomenclature, added the new properties (Vertices, Corners, NotCorners, Aeras, and AerasClock), as well as the new batching methods. These aren't too important anymore, since the old, complete implementation already allowed you to do the same thing. However, I felt that doing it as a Method rather than an operator was clearer for some, so I categorized it from Sugar and left it in the code.

The class was slightly reorganized and the documentation updated, and I realized how massive the work was.